Check Unstaged Files Using Grep
This began as a bugfix for `git commit --amend` commands that were
failing because `--amend` doesn't pass any files to pre-commit. This bug
is now fixed.

This also implements checking of all files. If staged files are passed
to pre-commit, TODOs in those files will block the commit. TODOs in
unstaged files are now also reported but will not cause the hook to
fail, resulting in a blocked commit due to files you haven't modified.

Because checking all files in a repo can be slow, we use grep to search
files in the repo for TODO items.